requirement.txt
assemblyai
openai==0.28
fastapi
requests
uvicorn
python-dotenv
python-multipart